Fix errors and get bot running with db refactor db refactor was successful. Plan on moving to different db other than sqlite. Probably use postgres db.
