+++ commands/popQuiz.js
@@ -3,12 +3,12 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { Sequelize, DataTypes, Op } = require('sequelize');
 const { MessageActionRow, MessageSelectMenu, Permissions, MessageEmbed } = require('discord.js');
-const quiz = require('./quiz.json');
+//const quiz = require('./quiz.json');
 
 module.exports = {
 	data: new SlashCommandBuilder()
-      .setName('popQuiz')
-      .setDescription('Pop quiz on a random topic');
+      .setName('popquiz')
+      .setDescription('Pop quiz on a random topic')
       .addSubcommand( subcommand =>
          subcommand
             .setName('start')
@@ -17,12 +17,12 @@ module.exports = {
                option
                   .setName('topic')
                   .setDescription('Specify the topic for the pop quiz')
-                  .setRequired(false)))
+                  .setRequired(false))),
    async execute(interaction, db){
       if(interaction.options.getSubcommand() === 'start'){
-         const quizObj = JSON.parse(quiz);
+         //const quizObj = JSON.parse(quiz);
          const topic = interaction.options.getString('topic');
-         const questions;
+         const questions = {};
          if(topic){
             //get questions from specified topic
             quizObj.topics.get(topic)
